refactor: move decoding logic in its own file
